empty-demo/empty-api/pyproject.toml
[project]
name = "empty-api"
version = "0.1.0"
description = "Simple FastAPI protected by Entra ID for OBO flow demo"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "pyjwt[crypto]>=2.9.0",
    "httpx>=0.28.0",
]
